+++ src/Core/Models/Api/Response/PlanResponseModel.cs
@@ -67,7 +67,7 @@ namespace Bit.Core.Models.Api
         public short? MaxUsers { get; set; }
 
         public bool HasAdditionalSeatsOption { get; set; }
-        public short? MaxAdditionalSeats { get; set; }
+        public int? MaxAdditionalSeats { get; set; }
         public bool HasAdditionalStorageOption { get; set; }
         public short? MaxAdditionalStorage { get; set; }
         public bool HasPremiumAccessOption { get; set; }
